+++ test/Blocks/Core.jl
@@ -8,7 +8,6 @@ using Yao
 @testset "parameters" begin
 g = chain(phase(0.1), phase(0.2), X, phase(0.3))
 
-@test parameters(g) isa Array{Float64, 1}
 @test nparameters(g) == 3
 
 for (i, each) in enumerate(parameters(g))
@@ -16,7 +15,6 @@ for (i, each) in enumerate(parameters(g))
 end
 
 g = kron(5, 3=>g, X, phase(0.4))
-@test parameters(g) isa Array{Float64, 1}
 @test nparameters(g) == 4
 
 for (i, each) in enumerate(parameters(g))
@@ -25,7 +23,6 @@ end
 
 g = rollrepeat(4, chain(phase(0.1), shift(0.2)))
 @test nparameters(g) == 8
-@test parameters(g) isa Array{Float64, 1}
 
 for (i, each) in enumerate(parameters(g))
     if i % 2 == 0